+++ src/calcTimes.ts
@@ -4,6 +4,8 @@
 
 import { formatTime } from "nrel-spa";
 import type { HighLatitudeRule } from "./highLatitude.js";
+import { defaultTimezone } from "./civilDate.js";
+import type { CivilDateInput } from "./civilDate.js";
 import { getTimes } from "./getTimes.js";
 import type { FormattedPrayerTimes } from "./types.js";
 
@@ -29,10 +31,10 @@ import type { FormattedPrayerTimes } from "./types.js";
  * console.log(times.Maghrib); // "20:31:17"
  */
 export function calcTimes(
-  date: Date,
+  date: CivilDateInput,
   lat: number,
   lng: number,
-  tz: number = -date.getTimezoneOffset() / 60,
+  tz: number = defaultTimezone(date),
   elevation = 0,
   temperature = 15,
   pressure = 1013.25,